+++ sanzu/src/client.rs
@@ -29,11 +29,14 @@ use crate::{
     //proto::{Tunnel, ReadWrite},
     sound::SoundDecoder,
     utils::{
-        get_xwd_data, ClientArgsConfig, HasTimeout, MAX_BYTES_PER_LINE, MAX_CURSOR_HEIGHT,
+        get_xwd_data, set_tcp_timeout, ClientArgsConfig, MAX_BYTES_PER_LINE, MAX_CURSOR_HEIGHT,
         MAX_CURSOR_WIDTH, MAX_WINDOW_HEIGHT, MAX_WINDOW_WIDTH,
     },
     video_decoder::init_video_codec,
 };
+
+#[cfg(target_family = "unix")]
+use crate::utils::HasTimeout;
 
 struct ShellAttr {
     path: &'static str,
@@ -230,9 +233,10 @@ pub fn do_run(
                 let destination = format!("{}:{}", arguments.server_addr, port);
                 let server = TcpStream::connect(&destination)
                     .context(format!("Error in tcp server connection {destination:?}"))?;
-                server
-                    .set_connection_timeout(connection_timeout)
-                    .context("Cannot set timeout")?;
+
+                let socket_ref = socket2::SockRef::from(&server);
+                set_tcp_timeout(socket_ref, connection_timeout).context("Cannot set keepalive")?;
+
                 info!("Connected to server");
                 server.set_nodelay(true).expect("set_nodelay call failed");
                 Box::new(server)
@@ -243,9 +247,8 @@ pub fn do_run(
                 let destination = format!("{}:{}", arguments.server_addr, port);
                 let server = TcpStream::connect(&destination)
                     .context(format!("Error in tcp server connection {destination:?}"))?;
-                server
-                    .set_connection_timeout(connection_timeout)
-                    .context("Cannot set timeout")?;
+                let socket_ref = socket2::SockRef::from(&server);
+                set_tcp_timeout(socket_ref, connection_timeout).context("Cannot set keepalive")?;
                 info!("Connected to server");
                 server.set_nodelay(true).expect("set_nodelay call failed");
                 Box::new(server)